refactor(download kit button): capital first letter
