+++ packages/gatsby-theme-styleguide/src/gatsby-components/DesignerKitLink/index.js
@@ -1,0 +1,1 @@
+export { default } from './DesignerKitLink'
