--- packages/gatsby-theme-styleguide/src/gatsby-components/designerKitLink/index.js
@@ -1,1 +1,0 @@
-export { default } from './designerKitLink'
